+++ pysmi/__init__.py
@@ -1,5 +1,5 @@
 # http://www.python.org/dev/peps/pep-0396/
-__version__ = '0.1.4'
+__version__ = '0.2.1'
 
 import sys